Rename: 공유방 관련 API 파일명 변경
공유관련 API 파일명 변경
